\name{TMBdebug-package}
\alias{TMBdebug-package}
\alias{TMBdebug}
\docType{package}
\title{
Debugger for TMB for windows machines
}
\description{
\packageDescription{TMBdebug}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{TMBdebug}
\packageIndices{TMBdebug}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{TMBdebug}

Maintainer: \packageMaintainer{TMBdebug}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
